+++ src/games/cclcc/tipssystem.cpp
@@ -15,7 +15,7 @@ using namespace Impacto::Io;
 
 void TipsSystem::DataInit(int scriptBufferId, uint8_t *tipsData,
                           uint32_t tipsDataSize) {
-  ScriptBufferId = scriptBufferId;
+  ScriptBufferId = (uint8_t)scriptBufferId;
   auto scriptBuffer = ScriptBuffers[scriptBufferId];
 
   TipEntryCount = 0;
@@ -30,9 +30,9 @@ void TipsSystem::DataInit(int scriptBufferId, uint8_t *tipsData,
     }
     // Read tip entry from the data array
     TipsDataRecord record;
-    record.Id = TipEntryCount;
+    record.Id = (uint16_t)TipEntryCount;
     // TODO: record.SortLetterIndex
-    record.NumberOfContentStrings = numberOfContentStrings;
+    record.NumberOfContentStrings = (uint16_t)numberOfContentStrings;
     for (int i = 0; i < numberOfContentStrings + 4; i++) {
       record.StringPtrs[i] =
           ScriptGetStrAddress(scriptBuffer, ReadLE<uint16_t>(&stream));
